requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.28
cohere==4.53
requests==2.31.0
langchain-community==0.0.28
pandas==2.2.1
tqdm==4.66.2